# backend/cozycoves/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/requests/update/{requestNo}": {
      "put": {
        "tags": [
          "request-controller"
        ],
        "operationId": "updateRequest",
        "parameters": [
          {
            "name": "requestNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/update/{houseNo}": {
      "put": {
        "tags": [
          "house-controller"
        ],
        "operationId": "updateHouse",
        "parameters": [
          {
            "name": "houseNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/House"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/requests/create": {
      "post": {
        "tags": [
          "request-controller"
        ],
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/create": {
      "post": {
        "tags": [
          "house-controller"
        ],
        "operationId": "createHouse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/House"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/House"
                }
              }
            }
          }
        }
      }
    },
    "/requests/byRenter/{username}": {
      "get": {
        "tags": [
          "request-controller"
        ],
        "operationId": "getRequestsByRenter",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/requests/byHouse/{houseNo}": {
      "get": {
        "tags": [
          "request-controller"
        ],
        "operationId": "getRequestsByHouse",
        "parameters": [
          {
            "name": "houseNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "getAllHouses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/House"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/houses/{house_no}": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "getAHouse",
        "parameters": [
          {
            "name": "house_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/search": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "searchHouses",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "bathrooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/byState/{state}": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "getHousesByState",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/byRenter/{username}": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "getHousesByRenter",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/byOwner/{username}": {
      "get": {
        "tags": [
          "house-controller"
        ],
        "operationId": "getHousesByOwner",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/requests/delete/{requestNo}": {
      "delete": {
        "tags": [
          "request-controller"
        ],
        "operationId": "deleteRequest",
        "parameters": [
          {
            "name": "requestNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/houses/delete/{houseNo}": {
      "delete": {
        "tags": [
          "house-controller"
        ],
        "operationId": "deleteHouse",
        "parameters": [
          {
            "name": "houseNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "House": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object"
          },
          "houseNo": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string"
          },
          "no_of_rooms": {
            "type": "integer",
            "format": "int32"
          },
          "no_of_bathrooms": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "renter": {
            "$ref": "#/components/schemas/User"
          },
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Request"
            }
          },
          "idAsString": {
            "type": "string"
          }
        }
      },
      "ObjectId": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Request": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "requestNo": {
            "type": "string"
          },
          "house": {
            "$ref": "#/components/schemas/House"
          },
          "requestedRenter": {
            "$ref": "#/components/schemas/User"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object"
          },
          "username": {
            "type": "string"
          }
        }
      }
    }
  }
}